VideoPriceCode.Regular.getRentCharge 부분 charge double로 변경

+++ VRUI/VideoPriceCode.java
@@ -1,9 +1,8 @@
 package VRUI;
-
 public enum VideoPriceCode {
     REGULAR {
         double getRentCharge (int daysRented) {
-            int charge = 2;
+            double charge = 2;
             if (daysRented > 2) charge += (daysRented-2)*1.5;
             return charge;
         }
